Rename package deviceplugin => devicemanager.
