+++ pkg/kubelet/cm/BUILD
@@ -123,7 +123,7 @@ go_library(
             "//pkg/apis/core/v1/helper:go_default_library",
             "//pkg/apis/core/v1/helper/qos:go_default_library",
             "//pkg/kubelet/cadvisor:go_default_library",
-            "//pkg/kubelet/cm/deviceplugin:go_default_library",
+            "//pkg/kubelet/cm/devicemanager:go_default_library",
             "//pkg/kubelet/cm/util:go_default_library",
             "//pkg/kubelet/events:go_default_library",
             "//pkg/kubelet/metrics:go_default_library",
@@ -219,7 +219,7 @@ filegroup(
         ":package-srcs",
         "//pkg/kubelet/cm/cpumanager:all-srcs",
         "//pkg/kubelet/cm/cpuset:all-srcs",
-        "//pkg/kubelet/cm/deviceplugin:all-srcs",
+        "//pkg/kubelet/cm/devicemanager:all-srcs",
         "//pkg/kubelet/cm/util:all-srcs",
     ],
     tags = ["automanaged"],
