+++ pkg/kubelet/cm/devicemanager/pod_devices.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package deviceplugin
+package devicemanager
 
 import (
 	"github.com/golang/glog"
